indexeddb: Report error message when transaction internally aborted
OnAbort was written with the assumption that it would only be called
from the front-end. This assumption is incorrect, since OnAbort can be
called from quota checks during commit. This results in the
out-of-quota error being dropped.

This CL adds a new kAborting state to distinguish aborts initiated
from the front-end from aborts initiated due to internal errors, such
as out-of-quota errors.

Bug: 1203335
Change-Id: I51c928e580fe8a059bfafa5cbb4714e6dd8759ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2857405
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Austin Sullivan <asully@chromium.org>
Cr-Commit-Position: refs/heads/master@{#881807}