[ozone] Fixed the drag and drop tests.
Two tests (see the linked crbug) did not work on Ozone/X11.  It turned
out that the origin of the window that initiated the drag could change
after its configuration.  On the other hand, the test always starts the
drag at position (0,0), which is outside the window if the one has
non-zero origin.

This CL fixes the issue by adjusting the size of the drag widget and the
start drag location so that drag happens within the window on all
existing Ozone platforms.  This approach is the simpler than originally
proposed in [1].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2875432

Bug: 1119787
Change-Id: Ic7f408c4b80b1ac0d801e6fdb0ba1d8f5b837b2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2885261
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Auto-Submit: Alexander Dunaev <adunaev@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#881843}