[View] Send gesture end events to the gesture handler to be replaced
In some scenarios such as showing the context menu, we need to reset
the gesture handler stored in the RootView instance. However, the
gesture handler replaced is not aware of the replacement and it may
expect to receive gesture end events.

This CL employs a two-step solution to fix this issue:

(1) Step 1: Check whether the gesture handler to be removed is under
any touches. If not, we don't need to send gesture end events;
otherwise, go to the step 2.

(2) Step 2: Send synthesized gesture end events to the target handler.

See this doc for further details: go/thoroughfix1163282

Bug: 1163282,1203812
Change-Id: Ia322488f3f328541e6e049a37ca135fc0f51b904
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2689036
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#881909}