VideoFrameYUVMailboxesHolder: Add SkImage and SkSurface methods
The one external caller to VideoFrameToSkiaTextures immediately
takes the resulting GrYUVABackendTextures and creates an SkImage
out of them. Make the VideoFrameToSkiaTextures method private, and
expose a new VideoFrameToSkImage method.

Expose a similar VideoFrameToPlaneSkSurfaces that creates SkSurfaces
for the individual planes.

SkSurfaces get a bit more complicated, because of the pixel format
used for the Y plane. In GLES 2.0, the format GL_LUMINANCE is the
canonical single channel format. GLES 3.0 added GL_RED (along with
GL_RG for the two channel format). Modern APIs (Metal, Vulkan, and
DXGI) don't expose LUMINANCE.

In Skia, LUMINANCE and kGray_8_SkColorType are not acceptable
formats for an SkSurface (but are acceptable for an SkImage). Add
a "for_surface" parameter in VideoFrameYUVMailboxesHolder to indicate
which type of import is desired.

Bug: 1206168
Change-Id: I2ffa9c8f5d17902e37dee6d620a005c2b36ae1ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2882138
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#881977}