Use the ComputeTilePhase() helper more in BackgroundImageGeometry
Although the expressions may appear different they are in fact computing
the same thing.

Add handling of zero tile extent to the helper. Also simplify it by
using the IntMod(...) function which computes the remainder in the
required way. The error/difference compared to the previous version
should be within LayoutUnit epsilon (and the new version is "LayoutUnit
exact").

Bug: 1208044
Change-Id: Ie8f8a514bccc7676b8be065b7a4d836961f038ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2883747
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#882075}