[profiles] Make kSessionExitedCleanly pref obsolete
kSessionExitedCleanly pref is deprecated but still checked in the code
as a fallback. The code contains a TODO to remove this pref around M28
release which makes me believe that this fallback is no longer needed.

The pref should be deleted from disk though, so this CL moves it
into the MigrateObsoleteProfilePrefs() function that takes care of it.

Change-Id: I84cddb655dd50d23578d4b9020d3efb9ca652ea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2891790
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882093}