[Web Payment] Measure install failure rate.
Before this patch, Chrome did not measure the failure rate for
installing a payment handler just-in-time, which made potential outages
difficult to detect.

This patch adds a PaymentRequest.PaymentHandlerInstallSuccess boolean
metric that is recorded when a just-in-time install of a payment handler
either succeeds or fails.

After this patch, Chrome measures the failure rate for installing a
payment handler just-in-time, which makes detecting potential outages
easier.

Bug: 1156875
Change-Id: If07b7ac731fb70d5519baf56b0a6b2b33833d335
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2888545
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Cr-Commit-Position: refs/heads/master@{#882105}