Privacy Sandbox Settings: Adjust FLoC logic to respect additional pref
A new preference to additionally control FLoC separately from the main
Privacy Sandbox control was recently added. This CL updates the logic
in the PrivacySandboxSettings service to additional respect the new
FLoC preference as an opt out.

FLoC was also relying on the IsPrivacySandboxAllowed function to check
whether it could run computation, adjusted this to use an updated
general IsFlocAllowed function, and renamed the existing FLoC function
to indicate that it is context specific.

Bug: 1206250
Change-Id: Icdf988d47230a5b48b0513064c40efe6d9756719
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2884159
Reviewed-by: Sean Harrison <harrisonsean@chromium.org>
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/master@{#882119}