[Nearby] Put Nearby Share eligibility for child accounts behind flag
Currently, Nearby Share if fully functional for child accounts. In
this CL, put child account eligibility behind an disabled-by-default
flag.

Manually verified that the Nearby Share service is not started for
child accounts if the flag is disabled, and verified that Nearby
Share is fully functional for child accounts when the flag is enabled.

Bug: 1206394
Change-Id: I5f944d977882be9a0463513acc0bdff6e74f6b73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2878646
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: James Vecore <vecore@google.com>
Cr-Commit-Position: refs/heads/master@{#882163}