[Autofill] Add StrikeDatabase to PersonalDataManager
This CL adds the strike database to the personal data manager which
counts the number of declines for the import of a new profile on a given
domain.
A simple interface is added to probe, add and clear the strikes.
Additionally, the history observer interface is used to clear the
strike records when the browsing history is cleared. For this, the
underlying database is changed to work with the observer paradigm.

Bug: 1175693
Change-Id: I3a65a00a7911ccda930e92125645103505648f89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2886900
Reviewed-by: Elizabeth Popova <lizapopova@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Matthias Körber <koerber@google.com>
Cr-Commit-Position: refs/heads/master@{#882190}