WebXR: update viewport scaling e2e tests
Add a frame count limit to the recommended-scale test to ensure
that the test fails cleanly if the feature isn't working as
expected, as opposed to timing out which looks too much like
a flaky test. (At this time testRecommendedViewportScale is
still disabled due to an underlying issue being investigated.)

Also add a bit of additional debugging information to help
show what the test was doing in case of failure. This status
is included in the test result string to avoid cluttering up the log.

Bug: 1207984
Change-Id: If91cce91cf2e4d4b86c762bc4450a32b7ec46c46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2892026
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Klaus Weidner <klausw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882234}