Fix use counter for broken overscroll-behavior.
The use counter we currently landed counts any time the document and
body style have a different overscroll-behavior, however, this only
matters when the body style was propagated to the viewport instead of
the document style. Otherwise, the behavior will not break when we
remove this body style propagation.

Bug: 954423
Change-Id: I2bac53097c44d2c7f25e587dd7e3cd45348c3c66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2892623
Commit-Queue: Robert Flack <flackr@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882238}