Canvas/WCG: Hide rec2020 behind a flag
We will need to have PredefinedColorSpace enum values hidden behind
runtime features. Right now, 'rec2020' is not officially supported but
is under development (and more are coming).

Add a function to validate a PredefinedColorSpace and throw the
appropriate exception depending on the presence or absence of runtime
features.

Change ToCanvasContextCreationAttributes, which converts
CanvasContextCreationAttributesModule to CanvasContextCreationAttributes,
to be able to fail and throw an exception (the HTML spec explicitly
indicates that "This can throw an exception" for the operation). Do so
on an invalid color space.

Bug: 1083693
Change-Id: I3f473cef65d0f0467e9673c10339bdc82aca262e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2891257
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882246}