Embed AR Content within the main Chrome hierarchy
Currently, the AR Surface view is loosely parented at the very top of
the App's hierarchy (on the 'content' element). This has been fine
previously, where we didn't want to enable anything else to show on top
of the AR Session, and also mimicked the previous Dialog behavior.
However, we *do* want to be able to display InfoBars and Permission
prompts on top of the AR Session. To this end, the AR Surface needs to
be embedded at a different point in the Main chrome hierarchy. With
this fix applied, infobars can start showing up over the AR content,
the same as if it were an ordinary fullscreened page. Future work will
enable permission prompts to appear over the AR content.

Unfortunately, due to the need to forward touch events to the DOM
content, we cannot parent the AR Surface to the compositor view holder,
as doing so results in an infinite loop where we forward events to the
compositor view holder, which forwards them back to us.

This change thus adds a new empty view holder target in front of the
compositor content, but behind everything else, as well as an
integration point for embedders to specify their own position, since
this is not a generic point (unlike the "content" point that was
used previously).

Bug: 1199437
Change-Id: I05639d5af8ae2b2efa35e1565844547f77600d79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2885322
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882250}