ClearOnExit: Continue unfinished deletions on startup
Sometimes Chrome is shut down before deletions for session-only
data are finished. This adds an additional cleanup step at startup
if it is detected that the deletion at shutdown did not finish.
This new behavior is behind a default-enabled Feature in case this is
causing any issues.

Bug: 637822, 1204022
Change-Id: Iebbf41a8f7f0b194563964f02df6ab4e0a824e66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2867585
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882258}