[Zenith] Pass StartPresentationContext to CastDialogController
When the Zenith dialog is opened from an in-page cast button, the
StartPresentationContext is not passed to the CastDialogController as it
does for the Cast dialog. As a result the callback
StartPresentationContext::HandleRouteResponse() is not run, and the
site does not know the result of a presentation request.

This CL initializes the CastDialogController using
|initWithStartPresentationContext()| and passes the ownership of the
object from the PresentationRequestNotificationItem to MediaRouterUI.

Bug: b/185135370, 1198234
Change-Id: Ie2e387a74edd1eec0982e2bdd5f5ca4167ed62d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2889497
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Commit-Queue: Muyao Xu <muyaoxu@google.com>
Cr-Commit-Position: refs/heads/master@{#882263}