[remoting][linux host] Fallback if remote-open-url isn't run on remote session
This CL makes the remote-open-url binary fallback to open the URL using
the previous default browser if the binary is not launched from the
remote session. This is needed since we can't guarantee that the default
browser settings will always be cleanly reset after the remote session
is ended.

Bug: b:183135000
Change-Id: If8334fc03e6c34c094b9429daf688ae6c3edc031
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2889519
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882267}