Reland "Fix threading issues in the OptGuideModelExecutor"
The first patchset is the original change, and the changes since are the
fix. In short, there's something on x86 devices that complicates
multi-thread testing where the second thread my never be started or
even deadlock. Thus, the solution is to provide the TaskRunner to the
executor handle so that it can be correctly given for testing on x86.

The failing devices are added as try bots.

This is a reland of 2c948437e5b26314f4e13bf1c7f8fd9eec9635c0

Original change's description:
> Fix threading issues in the OptGuideModelExecutor
>
> This CL splits the model executor into two parts, a UI part and a
> background thread counterpart. The UI class is called a UIHandle and
> the background class is the Executor.
>
> Robust sequence checking and guards have been put in place as well to
> aid in future development.
>
> See code comments or bug for more details.
>
> Bug: 1202253
> Change-Id: I3ab86d92eb288875841655203d34964b64c1c9de
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2861307
> Commit-Queue: Robert Ogden <robertogden@chromium.org>
> Reviewed-by: Michael Crouse <mcrouse@chromium.org>
> Reviewed-by: Sophie Chang <sophiechang@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#881221}

Bug: 1202253
Change-Id: I2af7954a20a495e5cf8f5b664588a521e8a8eafc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2888305
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Michael Crouse <mcrouse@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882277}