feed: ntp MV tiles should not see feed navigations
Clicking on NTP feed cards should not influence the MV
tiles on the NTP. The prior implementation relied on setting
a unique referrer. This was brittle, and finally broke when
the referrer of the feed was changed.

This CL instead remembers the last several Feed articles
tapped, and then checks this list to determine if the navigation
is from the Feed. This isn't perfect. If the user taps
on a Feed card, and then later navigates to that same page from
another method, we will ignore the navigation for MV tiles purposes
when we shouldn't

Bug: 1194788
Change-Id: I024e1829f5236f3b3566b0261e88e0475801b673
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2878557
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Commit-Queue: Dan H <harringtond@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882285}