Always notify FRE fragments that native init.
This is fixing a regression caused by https://crrev.com/c/2855038 that
will sometimes infinitely show a spinner when trying to accept the ToS.
The problem is we switched to using the ActivityLifecycleDispatcher to
notify fragments that native initialized. However, if native already
initialized, then registering for a native initialization signal never
sends a signal.

Work around this by explicitly checking if native has initialized and
manually notifying when in this scenario.

Bug: 1207683
Change-Id: Ib316d9dfb5355c7352ac7d88951e1f735aa55660
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2891917
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882286}