Update privacy handling of protected media identifier
This changes how we deal with the protected media identifier on
ChromeOS and Windows by no longer providing the ASK option and only
having ALLOW and BLOCK. This will not change functionality on Android.
The default will be ALLOW. The user still has the setting to change the
default to BLOCK.  Per-site settings will now be allowed regardless of
the default value (previously, they were only allowed when the default
was ASK and not BLOCK).

This migrates the existing kEnableDRM preference to now be the default
value for the protected media identifier content setting.

This will not change functionality on Android, only on ChromeOS and
Windows.

BUG=1191642,1085448
TEST=Unit tests pass, sites no longer prompt for protected content,
     allow/block behavior is consistent with design

Change-Id: Iab2599893dd357ad3f7e514abe29c18786f12175
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2488340
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882290}