Switch ALL_BRANCHES to LTS_MILESTONE where appropriate.
A number of CrOS LTS builders and consoles are marked with the
ALL_BRANCHES branch selector. This works fine when CrOS LTS is the only
type of branch besides standard, but will be incorrect when a branch
type for desktop extended stable is added.

Change-Id: Ic031c7a17b78f37ef87366d95ba3cf6f0f7a1ca6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2891491
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882342}