Reland: Log RECORD_AUDIO permission state on session start.
This was reverted due to tests failing on versions prior to Android M. The reason tests fail there is that WindowAndroid, which is mocked in the tests, have a dependency android.View.Display.Mode which is not supported in pre-M versions of Android. To fix, I've updated UmaSessionStats to take a (more appropriate) AndroidPermissionDelegate instead of a WindowAndroid. I've tested this locally on a Lollipop emulator and it has fixed the failures seen in crbug/1204549 👍

Original description:

This will be used to determine how often users are eligible for system voice search, including those that don't actively use it.

Bug: 1204310
Change-Id: I4a2b57fb77ed07c85e219fbd07714dc1a4ce1611
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2870009
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Josh Simmons <jds@google.com>
Cr-Commit-Position: refs/heads/master@{#882374}