[Live Caption] Send speech results directly from the speech service to the UI in the browser.
Before this change, speech results made the following journey:
1. Started in the speech service in the utility process on Lin/Mac/Win
   or the browser process on ChromeOS
2. Went to the renderer process
3. Went to the browser process

This change removes the additional hop to the renderer process (step 2)
and has results go directly from the speech service to the UI in the
browser process (steps 1 and 3).

This does not have any user-facing change other than being a slight
optimization. The main motivation is to prepare the speech service to
accept inputs from multiple sources, not just the renderer process in
the browser.

For a more detailed design, see the before/after diagrams here:
https://docs.google.com/document/d/1Janw-wF2oWrqK0-KTk4433wu4x_QeciW_0nnSkxBShw/edit?pli=1#heading=h.vy3wyozaqxxc

Change-Id: I982e407f673934194ec48d267765f8bbaadd8f9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2826803
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Rob Schonberger <robsc@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Evan Liu <evliu@google.com>
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/master@{#882427}