Add mwc-button to Sample System Web App
This CL adds an mwc-button to the Chrome OS Sample System Web App and
adds the build pipeline. The build_mwc_app.py script is largely a copy
of the optimize_webui pipeline. Future CLs will prune down and remove
unnecessary parts, but we intend to land as-is so that we can begin
integration prototyping.

Bug: b/176932745

Change-Id: I39ba812f07beb19281df24b8ee5bfa598b9a7859
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2775787
Commit-Queue: calamity <calamity@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Jiewei Qian  <qjw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882454}