PowerMode: Instrument in-flight resource loads for PowerMode::kLoading
Add a PowerModeVoter to ResourceLoadObserverForFrame, which votes for
kLoading mode as long as at least 3 pending requests exist.

This widens the period during which we detect kLoading PowerMode from
just the per-frame onLoad signal (instrumented in FrameSchedulerImpl)
to periods with a non-insignificant amount of network activity and also
adds a 1 second timeout after loading.

Bug: 1166695
Bug: b:187930150
Change-Id: I3efb0eee4f93c540f389c378a951e5e94814af3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2888165
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882499}