Revert "views: support Mac textfield-focus behavior"
This reverts commit fa2d40dd48636a51389970a21eb1dce68dd506b0.

kTextfieldScrollsToStartOnFocusChange was well-intentioned, but it
was not implemented with all the subtleties that would be required
for a full re-creation of the AppKit behavior. Implementing all
those subtleties would require a lot more work, plus without those
subtleties, kTextfieldScrollsToStartOnFocusChange causes issues.
See https://crbug.com/1208119 for an example of the issues and for
more details on why this needs to be done.

Fixed: 1208119
Fixed: 1206675
Bug: 657558
Change-Id: I84e0567efa555309de65aa44d2e7cde71e1ed79f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2891442
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882510}